Fix for second audio recording block
The audio recording was not working on the second block as we were re-using the same id.  Now every time we start a new recording we will ensure that the id is regenerated.

Also there is a technical detail about the buffer size.  Instead of passing in a fixed buffer size -  we can send in 0 and the os should pick the best buffer size for createScriptProcessor which is used in the microphone level processing.